requirements.txt
annotated-types==0.7.0
anthropic==0.42.0
anyio==4.8.0
certifi==2024.12.14
contourpy==1.3.1
cycler==0.12.1
distro==1.9.0
fonttools==4.55.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
joblib==1.4.2
kiwisolver==1.4.8
matplotlib==3.8.2
numpy==1.26.3
packaging==24.2
pandas==2.2.0
pillow==11.1.0
pyarrow==18.1.0
pydantic==2.10.5
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
scikit-learn==1.4.0
scipy==1.15.1
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
tabulate==0.9.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
